Move controller functions 	- Creates new controllers that were needed 	- Moves all functions to their appropriate controllers 	- Converts some controllers from RestControllers to Controllers
