+++ src/main/java/ShopifyProj/HomeController.java
@@ -1,0 +1,26 @@
+package ShopifyProj;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.*;
+
+import javax.persistence.Id;
+import javax.swing.text.html.Option;
+import java.text.NumberFormat;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String viewHomePage(Model model) {
+        return "homepage";
+    }
+
+}
