+++ src/main/java/ShopifyProj/ShopController.java
@@ -1,10 +1,9 @@
 package ShopifyProj;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.HashSet;
 import java.util.List;
@@ -13,10 +12,38 @@ import java.util.Set;
 import java.util.ArrayList;
 import java.util.concurrent.atomic.AtomicLong;
 
-@RestController
+@Controller
 public class ShopController {
     @Autowired
     private ShopRepository shopRepo;
+
+    private Shop getShopById(int aShopId) {
+        return shopRepo.findById(aShopId);
+    }
+
+    private List<Shop> getShops() {
+        List<Shop> shops = new ArrayList<Shop>();
+        for (Shop shop : shopRepo.findAll()) {
+            shops.add(shop);
+        }
+
+        return shops;
+
+    }
+
+    @GetMapping("/goToAddShopPage")
+    public String viewAddShopPage(Model model) {
+        model.addAttribute("shops", getShops());
+        model.addAttribute("shop", new Shop());
+
+        return "addShopPage";
+    }
+
+    @GetMapping("/goToShop")
+    public String viewShopPageById(@RequestParam(value = "shopId") int aShopId, Model model) {
+        model.addAttribute("shop", getShopById(aShopId));
+        return "shopPage";
+    }
 
     @PostMapping("/addShop")
     public @ResponseBody Shop addShop(@RequestParam(value = "shopName") String name,
@@ -38,25 +65,22 @@ public class ShopController {
         return newShop;
     }
 
-    @PostMapping("/search")
-    public @ResponseBody ArrayList<Shop> search(@RequestParam(value = "searchField") String query) {
-        ArrayList<Shop> matchingShops = new ArrayList<>();
-        for (Shop shop : shopRepo.findAll()) {
-            boolean isAdded = false;
+    @GetMapping("/YourShopPage")
+    public String displayYourShop(@RequestParam(value = "shopId") Integer shopId, Model model){
 
-            Set<Tag> tags = shop.getTags();
-            for (Tag t : tags) {
-                if(t.getTagName().equals(query) || t.getTagName().contains(query)) {
-                    matchingShops.add(shop);
-                    isAdded = true;
-                }
-            }
+        Shop toView = null;
 
-            if (shop.getShopName().equals(query) && isAdded == false) {
-                matchingShops.add(shop);
-            }
+        Optional<Shop> theShop = shopRepo.findById(shopId);
+        if(theShop.isPresent()){
+            toView = theShop.get();
+            System.out.println(toView);
+        }else {
+            System.out.println(String.format("No shop found with ID %d", shopId));
         }
-        return matchingShops;
-    }
 
+        model.addAttribute("shop", toView);
+        model.addAttribute("item", new Item());
+
+        return "MerchantShopPage";
+    }
 }
